Structral changes. Readme updated (Not done yet).
