+++ README.md
@@ -1,10 +1,13 @@
 # Dota2 Win Probability Prediction Using Replay Files
 
-This project trained and provided models to predict and plot the win probability of both teams over time, given the replay file of a specific game. The example graph is shown below.
+# Table of Contents
+1. [Motivations](README.md#motivations)
+2. [Overview of the Project](README.md#overview-of-the-project)
+3. [Repo Structure](README.md#repo-structure)
+9. [Setup](README.md#setup)
+10. [Usage](README.md#usage)
 
-<p float="left">
-  <img src="docs/images/win_prob_demo.gif" width="800" />
-</p>
+This project trained and provided models to predict and plot the win probability of both teams over time, given the replay file of a specific game.
 
 To check-out the repo:
 
@@ -12,23 +15,72 @@ To check-out the repo:
 git clone https://github.com/henryhao1991/Dota2-Win-Probability-Prediction.git
 ```
 
-## Motivation for the Project
+## Motivations
 
 E-sports is a rapidly growing field and draws a lot of public attention this year. Dota2 has a subscription service called DotaPlus, including a feature that shows a real time win probability graph, but its algorithm is not revealed. I would like to build a model that train from the replay files to do the similar task: given the time and game information, predict the win probability of both teams.
 
 ## Overview of the Project
 
-In this project, I trained LSTM models that takes the processed inputs from a game replay, and predict the win probability of Radiant side (In Dota2, players joins two teams, Radiant and Dire).
+In this project, I trained LSTM models that takes the processed inputs from a game replay, and predict the win probability of Radiant side (In Dota2, players joins two teams, Radiant and Dire). The details of the implementations (data processing, models, etc.) are provided at the end.
 
-Details of the implementation:
+The deliverable of the project is a Streamlit App that allows user to specify a replay file, and then parses the file and generates the win probability graph using the selected model. The example of how the App looks like is shown below.
 
-- Raw data: Replay files (.dem format). Fetched using OpenDota's API.
+<p float="left">
+  <img src="docs/images/win_prob_demo.gif" width="800" />
+</p>
 
-- Data processing pipeline:
-  * Parse the replay file using a script provided by OpenDota
+## Repo Structure
 
-## Requirements
+Below is the structure of this repo.
+    ├── README.md
+    ├── images
+    │   └── Images used for readme
+    ├── model
+    │   └── __init__.py
+    │   └── model.py
+    │   └── hero2vec.py
+    ├── util
+    │   └── __init__.py
+    │   └── dataloader.py
+    │   └── data_process.py
+    ├── h2v_training.py
+    ├── training.py
+    ├── streamlit_demo.py
+    ├── saved_model
+    │   └── hero_embeddings.txt    
+    │   └── model_agg.pt
+    │   └── model_indi.pt
+    │   └── model_agg_h2v_subnet.pt
+    │   └── model_indi_h2v_subnet.pt
+    └── results_notebooks
+        └── hero2vec_result.ipynb
+        └── LSTM_training.ipynb
+        └── results_comparison.ipynb
 
 ## Run A Train and Test
 
-## Results
+## Details of the Implementation
+
+- Raw data: Replay files (.dem format). Fetched using OpenDota's API.
+<br/>
+- Data processing pipeline:
+  * Parse the replay file using a script provided by OpenDota:
+    https://github.com/odota/parser
+  * Get the game information every 30 second from the parsed JSON file.
+  * Extract useful information from the 30-second-interval entries.
+
+  All the data processing funtion can be found in data_process.py
+<br/>
+- Selected features (Used either team features or individual features):
+  * Team features: Team total gold and experience, number of players died on both teams, number of towers destroyed on both teams
+  * Individual features: Individual gold and experience, number of players died on both teams, number of towers destroyed on both teams
+  * Team embeddings: An vector embedding for heroes is trained. The graph below shows the results of the embedding in 2 eigenvector directions corresponding to the 2 largest eigenvalues of PCA. We can see, the results clustered based on the preferred roles for different heros (carries, mid-lane, off-lane, supports), while the more flexible heroes lies in the middle of their usual roles.
+<br/>
+  <p float="left">
+    <img src="docs/images/hero2vec.png" width="800" />
+  </p>
+
+- Model used:
+  * Heuristic model: Just a baseline model to compare with. Simply return the output of a Sigmoid function over the total difference of gold and experience between two teams dividing by a scaling factor.
+  * LSTM model: A single layer LSTM network, with 50 hidden nodes.
+  * LSTM model with subnet for team embeddings: The LSTM network plus a feed forward subnet with the team embeddings as the input. The results of both network are concatenated and are passed to a fully connected layer.
